% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_log.R
\name{average_log}
\alias{average_log}
\title{average_log}
\usage{
average_log(data, limit = 0.001, max_iter = 20)
}
\arguments{
\item{data}{data.table}

\item{limit}{numeric}

\item{max_iter}{numeric}
}
\value{
list
}
\description{
Implement the average log algorithm.
}
\examples{
data = data.frame(source = c('source1','source1','source2','source2','source3','source3','source4',
'source4','source4'),
object = c('x','y','y','z','x','z','x','y','z'),
fact = c('4','7','7','5','3','5','3', '6','8'))
average_log(data)
}
