% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_belief.R
\name{get_prior_belief}
\alias{get_prior_belief}
\title{get_prior_belief}
\usage{
get_prior_belief(data, type)
}
\arguments{
\item{data}{data.table}

\item{type}{string}
}
\value{

}
\description{
Auxiliary function to produce the prio beliefs selected by the user on each algorithm.
Algorithms do have by default some prior belief
}
