% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investment.R
\name{investment}
\alias{investment}
\title{investment
Implement investment algorithm.  In the Investment algorithm, sources “invest”
their trustworthiness uniformly among their claims. The belief in each claim is
a non-linear function of g, normally 1.2, and the source’s trustworthiness is
updated as the sum of the beliefs in their claims, weighted by the proportion
of trust invested by all sources in the claim.
Since claims with higher-trust sources get higher belief, these claims become
relatively more believed and their sources become more trusted.}
\usage{
investment(data, limit = 0.001, g = 1.2, max_iter = 20)
}
\arguments{
\item{data}{data.table}

\item{limit}{numeric}

\item{g}{numeric}

\item{max_iter}{numeric}
}
\value{
list
}
\description{
investment
Implement investment algorithm.  In the Investment algorithm, sources “invest”
their trustworthiness uniformly among their claims. The belief in each claim is
a non-linear function of g, normally 1.2, and the source’s trustworthiness is
updated as the sum of the beliefs in their claims, weighted by the proportion
of trust invested by all sources in the claim.
Since claims with higher-trust sources get higher belief, these claims become
relatively more believed and their sources become more trusted.
}
\examples{
data = data.frame(source = c('source1','source1','source2','source2','source3','source3','source4',
'source4','source4'),
object = c('x','y','y','z','x','z','x','y','z'),
fact = c('4','7','7','5','3','5','3', '6','8'))
investment(data)
}
