% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_investment.R
\name{pooled_investment}
\alias{pooled_investment}
\title{Pooled investment algorithm}
\usage{
pooled_investment(data, limit = 0.001, g = 1.4, max_iter = 20)
}
\arguments{
\item{max_iter}{}
}
\value{

}
\description{
Pooled investment algorithm
}
