% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthfinder.R
\name{truthfinder}
\alias{truthfinder}
\title{Title}
\usage{
truthfinder(
  data,
  limit = 0.001,
  max_iter = 20,
  damping_factor = 0.3,
  influence_parameter = 0.5,
  initial_trust = 0.9,
  similarity = NULL
)
}
\arguments{
\item{similarity}{}
}
\value{

}
\description{
Title
}
